Added gather argument to parse, incremented version number.
In certain scenarios, especially during the course of working with `AIORcon`, it has become apparent that gathering all coroutines may not work if other libraries do not properly handle asyncio. As such, to help with avoiding this error a gather argument has been added to parse to act as a switch on whether or not to concurrently run all coroutines found or to await them individually.